Split oversized CI-gated modules
